+++ examples/data-sources/datahub_roles/data-source.tf
@@ -1,0 +1,5 @@
+data "datahub_roles" "all" {}
+
+output "role_urns" {
+  value = data.datahub_roles.all.urns
+}